// resources/js/api/generated/models/getApiV1ServersServeridStacksStacknameFiles200EntriesItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV1ServersServeridStacksStacknameFiles200EntriesItem = {
  extension?: string;
  group?: string;
  /** @minimum 0 */
  group_id?: number;
  is_directory: boolean;
  mod_time: string;
  mode: string;
  name: string;
  owner?: string;
  /** @minimum 0 */
  owner_id?: number;
  path: string;
  size: number;
};